refactoring behaviors
