+++ src/rer/ufficiostampa/behaviors/metadata.py
@@ -6,12 +6,28 @@ from plone.autoform import directives as form
 from plone.autoform.interfaces import IFormFieldProvider
 from zope import schema
 from zope.interface import provider
+from plone.app.z3cform.widget import AjaxSelectFieldWidget
 
 
 @provider(IFormFieldProvider)
 class IBasicComunicati(IBasic):
     title = schema.Text(title=_("label_title", default="Title"), required=True)
     form.widget("title", rows=2)
+
+    arguments = schema.Tuple(
+        title=_("arguments_label", default="Arguments"),
+        description=_("arguments_help", default="Select one or more values."),
+        value_type=schema.TextLine(),
+        required=True,
+        missing_value=(),
+    )
+
+    form.widget(
+        "arguments",
+        AjaxSelectFieldWidget,
+        vocabulary="rer.ufficiostampa.vocabularies.arguments",
+        pattern_options={"allowNewItems": "false"},
+    )
 
     searchable("title")
