+++ src/rer/ufficiostampa/interfaces/content.py
@@ -1,37 +1,12 @@
-from plone.app.contenttypes.behaviors.richtext import IRichText
-from plone.app.dexterity.textindexer import searchable
-from plone.app.z3cform.widget import AjaxSelectFieldWidget
-from plone.autoform import directives
+# from plone.app.contenttypes.behaviors.richtext import IRichText
+# from plone.app.dexterity.textindexer import searchable
+from plone.autoform import directives as form
 from plone.supermodel import model
 from rer.ufficiostampa import _
-from rer.ufficiostampa.utils import defaultLegislature
 from zope import schema
 
 
 class IComunicatoStampa(model.Schema):
-    arguments = schema.Tuple(
-        title=_("arguments_label", default="Arguments"),
-        description=_("arguments_help", default="Select one or more values."),
-        value_type=schema.TextLine(),
-        required=True,
-        missing_value=(),
-    )
-
-    directives.widget(
-        "arguments",
-        AjaxSelectFieldWidget,
-        vocabulary="rer.ufficiostampa.vocabularies.arguments",
-        pattern_options={"allowNewItems": "false"},
-    )
-
-    legislature = schema.TextLine(
-        title=_("label_legislature", default="Legislature"),
-        description="",
-        required=True,
-        defaultFactory=defaultLegislature,
-    )
-    directives.mode(legislature="display")
-
     message_sent = schema.Bool(
         title=_("label_sent", default="Sent"),
         description="",
@@ -40,11 +15,19 @@ class IComunicatoStampa(model.Schema):
     )
     comunicato_number = schema.TextLine(title="", description="", required=False)
 
-    directives.omitted("message_sent")
-    directives.omitted("comunicato_number")
+    tipologia_news = schema.TextLine(
+        title=_("label_tipologia_news", default="Tipologia News"),
+        description="",
+        required=True,
+        default="Comunicato Stampa",
+        readonly=True,
+    )
+    form.mode(tipologia_news="hidden")
+    form.omitted("message_sent")
+    form.omitted("comunicato_number")
 
     # set text field as searchable in SearchableText
-    searchable(IRichText, "text")
+    # searchable(IRichText, "text")
 
 
 class IInvitoStampa(IComunicatoStampa):
