+++ src/rer/ufficiostampa/profiles/default/types/ComunicatoStampa.xml
@@ -43,8 +43,26 @@
     <element value="plone.richtext" />
     <element value="plone.versioning" />
     <element value="plone.locking" />
-    <element value="rer.solrpush.behaviors.solr_fields.ISolrFields" />
     <element value="plone.textindexer" />
+
+    <!-- TODO: su profilo rer aggiungere
+      <element value="rer.solrpush.behaviors.solr_fields.ISolrFields" /> 
+      <element value="rer.ufficiostampa.legislature" />
+    -->
+
+    <!-- TODO:
+    nel policy per integrazione con io-comune
+
+    togliere:
+    <element value="rer.ufficiostampa.basic" />
+    <element value="rer.ufficiostampa.legislature" />
+
+    aggiungere:
+    <element value="plone.basic" />
+    <element value="design.plone.contenttypes.behavior.news_base" />
+    <element value="design.plone.contenttypes.behavior.argomenti_news" /> 
+    -->
+
   </property>
   <property name="schema">rer.ufficiostampa.interfaces.IComunicatoStampa</property>
   <property name="model_source" />
